Use C_INT64_T instead of C_LONG_LONG

+++ include/galahad_arc.h
@@ -211,6 +211,7 @@
 extern "C" {
 #else
 #include <stdbool.h>
+#include <stdint.h>
 #endif
 
 // include guard
@@ -590,7 +591,7 @@ struct arc_inform_type {
 
     /// \brief
     /// the maximum number of entries in the factors
-    int max_entries_factors;
+    int64_t max_entries_factors;
 
     /// \brief
     /// the total integer workspace required for the factorization
